Single method to build and inject SVG paths from layer
