+++ index.html
@@ -16,7 +16,7 @@
     <main>
       <h1 id="message"></h1>
       <p id="content"></p>
-      <div id="display"></div>
+      <svg id="display"></svg>
     </main>
   </body>
